requirements.txt
fastapi==0.111.0
fastapi-mail==1.4.1
PyJWT==2.8.0
uvicorn==0.30.1
passlib==1.7.4
python-jose==3.3.0
pydantic-settings==2.3.3
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
cachetools==5.3.3